Update to move writeXXXX functions to own file

+++ System/geometry/Mesh3D.cxx
@@ -3,7 +3,7 @@
  
  * File:   geometry/Mesh3D.cxx
  *
- * Copyright (c) 2004-2016 by Stuart Ansell
+ * Copyright (c) 2004-2018 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -45,6 +45,7 @@
 #include "BaseVisit.h"
 #include "BaseModVisit.h" 
 #include "support.h"
+#include "writeSupport.h"
 #include "MatrixBase.h"
 #include "Matrix.h"
 #include "Vec3D.h"
@@ -202,7 +203,7 @@ Mesh3D::getCoordinate(const std::vector<double>& Vec,
          
   
   return Vec[I]+static_cast<double>(Index-offset)*
-    (Vec[I+1]-Vec[I])/NF[I];
+    (Vec[I+1]-Vec[I])/static_cast<double>(NF[I]);
 }
 
 std::vector<Geometry::Vec3D>
